+++ src/main/kotlin/kotools/assert/Failure.kt
@@ -1,0 +1,24 @@
+package kotools.assert
+
+import org.junit.jupiter.api.assertDoesNotThrow
+
+/**
+ * Asserts that the [block] function doesn't throw an error and returns its
+ * result.
+ */
+@SinceKotoolsAssert("2.1")
+public fun <T> assertPass(block: () -> T): T = assertDoesNotThrow(block)
+
+/** Asserts that the [block] function throws an error and returns this error. */
+@SinceKotoolsAssert("2.1")
+public inline fun assertFails(block: () -> Unit): Throwable =
+    kotlin.test.assertFails(block)
+
+/**
+ * Asserts that the [block] function throws an error of type [T], and returns
+ * this error.
+ */
+@SinceKotoolsAssert("2.1")
+public inline fun <reified T : Throwable> assertFailsWith(
+    block: () -> Unit
+): T = kotlin.test.assertFailsWith(block = block)